@@ -19,8 +19,8 @@
     "build:js": "tsc --outDir dist/js -t ES2015",
     "build:es": "tsc --outDir dist/es -t ES2015 --module ES2015",
     "build:next": "tsc --outDir dist/esnext --module esnext",
-    "build:umd": "bili",
-    "build": "npm run checks && npm run build:js && npm run build:es && npm run build:next && npm run build:umd",
+    "build:umd": "vite build",
+    "build": "npm run checks && npm run build:umd && npm run build:js && npm run build:es && npm run build:next",
     "dev": "tsc --noEmit -w",
     "checks": "madge --circular --extensions ts src/index.ts",
     "commit-build": "(git diff --quiet && git diff --staged --quiet) || (git commit -am \"Update dist\")",
@@ -66,24 +66,26 @@
   "bugs": {
     "url": "https://github.com/Raiondesu/vuse-rx/issues"
   },
-  "homepage": "https://vuse-rx.raiondesu.rocks",
+  "homepage": "https://vuse-rx.raiondesu.dev",
   "devDependencies": {
-    "@types/jest": "^26.0.20",
-    "@types/node": "^12.11.7",
-    "bili": "^5.0.5",
-    "eslint": "^6.7.2",
-    "eslint-plugin-import": "^2.20.2",
+    "@types/jest": "^29.2.5",
+    "@types/node": "^18.11.18",
+    "eslint": "^8.32.0",
+    "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": "^4.0.0",
-    "jest": "^26.6.3",
-    "madge": "^4.0.2",
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-standard": "^5.0.0",
+    "jest": "^29.3.1",
+    "jest-environment-jsdom": "^29.3.1",
+    "madge": "^5.0.1",
     "rollup-plugin-typescript2": "^0.29.0",
-    "rxjs": "^7.1.0",
-    "ts-jest": "^26.5.0",
-    "typescript": "^4.2.3",
-    "vitepress": "^0.12.2",
-    "vue": "^3.0.5",
+    "rxjs": "^7.8.0",
+    "sitemap-ts": "^1.2.2",
+    "ts-jest": "^29.0.5",
+    "typescript": "^4.9.4",
+    "vite": "^4.0.4",
+    "vitepress": "^1.0.0-alpha.38",
+    "vue": "^3.2.45",
     "vuse-rx": "./"
   },
   "peerDependencies": {
@@ -93,6 +95,7 @@
   "jest": {
     "preset": "ts-jest",
     "verbose": true,
+    "testEnvironment": "jest-environment-jsdom",
     "coverageReporters": [
       "lcov",
       "text"
