@@ -34,7 +34,7 @@ export function syncRef<R1, R2>(
  * The resulting ref serves as an origin point for the binding,\
  * values **from** the resulting ref and **to** the resulting ref are mapped onto the first.
  */
- export function syncRef<R1, R2, M extends Readonly<Mappers<R1, R2>> = Mappers<R1, R2>>(
+export function syncRef<R1, R2, M extends Readonly<Mappers<R1, R2>> = Mappers<R1, R2>>(
   ref1: Ref<R1>,
   { to, from }: M,
   defaultValue?: R2,
