@@ -1,5 +1,3 @@
-import { Ref, UnwrapRef } from 'vue';
-
 export type Builtin =
   | Function
   | Date
@@ -13,8 +11,6 @@ export type Builtin =
   | symbol
   | undefined
   | null;
-
-export type UnwrapNestedRefs<T> = T extends Ref ? T : UnwrapRef<T>;
 
 export type MutationStrategy<S extends Record<PropertyKey, any>, M> = {
   /**
@@ -35,7 +31,7 @@ export type MutationStrategy<S extends Record<PropertyKey, any>, M> = {
  * @param mutation - the main checking reference
  * @param key - a key into which to advance
  */
- export const canMergeDeep = <S extends Record<PropertyKey, any>, Mutation extends Record<keyof S, any>>(
+export const canMergeDeep = <S extends Record<PropertyKey, any>, Mutation extends Record<keyof S, any>>(
   state: S,
   mutation: Mutation | null | undefined,
   key: keyof S,
